refactor: optimize LockmanResult Equatable usage and remove redundant conformance
- Add private Array<LockmanResult>.allSuccessful extension to LockmanCompositeStrategy
- Replace allSatisfy({ /bin/zsh == .success }) with more efficient pattern matching
- Remove public Equatable Conformance from LockmanResult (unused in production)
- Move Equatable Conformance to test-only helper for XCTAssertEqual support
- Improve performance by avoiding localizedDescription comparisons
- Clarify intent with dedicated allSuccessful property
