+++ Sources/Lockman/Core/Internal/LockmanResult.swift
@@ -46,25 +46,3 @@ public enum LockmanResult: Sendable {
   ///   and the new action was cancelled.
   case cancel(any LockmanError)
 }
-
-// MARK: - Equatable Conformance
-
-extension LockmanResult: Equatable {
-  public static func == (lhs: LockmanResult, rhs: LockmanResult) -> Bool {
-    switch (lhs, rhs) {
-    case (.success, .success):
-      return true
-    case (
-      .successWithPrecedingCancellation(let lhsError),
-      .successWithPrecedingCancellation(let rhsError)
-    ):
-      // Compare errors by their localized description since Error is not Equatable
-      return lhsError.localizedDescription == rhsError.localizedDescription
-    case (.cancel(let lhsError), .cancel(let rhsError)):
-      // Compare errors by their localized description since Error is not Equatable
-      return lhsError.localizedDescription == rhsError.localizedDescription
-    default:
-      return false
-    }
-  }
-}
